+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `[usage/completion/bash]` Added bash completion generator
 * `[usage/completion/zsh]` Added zsh completion generator
+* `[usage/completion/fish]` Added fish completion generator
 * `[usage]` Added method `info.BoundOptions` for linking command with options
 * `[csv]` Added method `reader.ReadTo` for reading CSV data into slice
 * `[strutil]` Fixed bug in `Fields` method